divide handlers into 3 files by purpose for clarity
