+++ GSP_API/runserver.py
@@ -2,23 +2,20 @@ import logging
 import sys
 from os import getenv
 
-from api_handlers import (forecast_stats_handler,
-                          forecast_ensembles_handler,
-                          forecast_warnings_handler,
-                          forecast_records_handler,
-                          historic_data_handler,
-                          return_periods_handler,
-                          seasonal_average_handler,
-                          get_available_data_handler,
-                          get_region_handler,
-                          get_dates_handler,
-                          get_reach_id_from_latlon_handler,)
 from flask import Flask, request, jsonify
-from flask_restful import Api
-
 from flask_cors import CORS, cross_origin
-
-blob_mapped_dir = "/mnt/output"
+from flask_restful import Api
+from handlers_forecasts import (forecast_stats_handler,
+                                forecast_ensembles_handler,
+                                forecast_warnings_handler,
+                                forecast_records_handler,
+                                available_dates_handler, )
+from handlers_historical import (historic_data_handler,
+                                 seasonal_average_handler,
+                                 return_periods_handler, )
+from handlers_utilities import (get_available_data_handler,
+                                get_region_handler,
+                                get_reach_id_from_latlon_handler, )
 
 print("Creating Application")
 
@@ -54,6 +51,7 @@ def init_logger():
 def forecast_stats():
     init_logger()
 
+    # ensure you have enough information provided to return a response
     if request.args.get('reach_id', '') == '':
         if request.args.get('lat', '') == '' or request.args.get('lon', '') == '' or \
                 request.args.get('region', '') == '':
@@ -61,9 +59,7 @@ def forecast_stats():
             return jsonify({"error": "Either reach_id or lat and lon parameters are required as input."}), 422
 
     try:
-        # Call the service
         return forecast_stats_handler(request)
-
     except Exception:
         print(sys.exc_info()[0])
         logging.exception(sys.exc_info()[0])
@@ -76,37 +72,34 @@ def forecast_stats():
 def forecast_ensembles():
     init_logger()
 
+    # ensure you have enough information provided to return a response
     if request.args.get('reach_id', '') == '':
         if request.args.get('lat', '') == '' or request.args.get('lon', '') == '':
             logging.error("Either reach_id or lat and lon parameters are required as input.")
             return jsonify({"error": "Either reach_id or lat and lon parameters are required as input."}), 422
 
     try:
-        # Call the service
         return forecast_ensembles_handler(request)
-
     except:
         print(sys.exc_info()[0])
         logging.exception(sys.exc_info()[0])
         return jsonify({"error": "An unexpected error occured."}), 400
 
 
-# GET, API ForecastEnsembles endpoint
+# GET, API ForecastWarnings endpoint
 @app.route(api_prefix + '/ForecastWarnings/', methods=['GET'])
 @cross_origin()
 def forecast_warnings():
     init_logger()
 
-    # ensure you have enough information provided to provide a response
+    # ensure you have enough information provided to return a response
     if not request.args.get('region', False):
         if not request.args.get('lat', False) or not request.args.get('lon', False):
             logging.error("Either reach_id or lat and lon parameters are required as input.")
             return jsonify({"error": "Either region or lat and lon parameters are required as input."}), 422
 
     try:
-        # collect the
         return forecast_warnings_handler(request)
-
     except:
         print(sys.exc_info()[0])
         logging.exception(sys.exc_info()[0])
@@ -119,16 +112,14 @@ def forecast_warnings():
 def forecast_records():
     init_logger()
 
-    # ensure you have enough information provided to provide a response
+    # ensure you have enough information provided to return a response
     if request.args.get('reach_id', '') == '':
         if request.args.get('lat', '') == '' or request.args.get('lon', '') == '':
             logging.error("Either reach_id or lat and lon parameters are required as input.")
             return jsonify({"error": "Either reach_id or lat and lon parameters are required as input."}), 422
 
     try:
-        # collect the
         return forecast_records_handler(request)
-
     except:
         print(sys.exc_info()[0])
         logging.exception(sys.exc_info()[0])
@@ -141,6 +132,7 @@ def forecast_records():
 def historic_simulation():
     init_logger()
 
+    # ensure you have enough information provided to return a response
     if request.args.get('reach_id', '') == '':
         if request.args.get('lat', '') == '' or request.args.get('lon', '') == '' or \
                 request.args.get('region', '') == '':
@@ -148,9 +140,7 @@ def historic_simulation():
             return jsonify({"error": "Either reach_id or lat and lon parameters are required as input."}), 422
 
     try:
-        # Call the service
         return historic_data_handler(request)
-
     except:
         print(sys.exc_info()[0])
         logging.exception(sys.exc_info()[0])
@@ -163,6 +153,7 @@ def historic_simulation():
 def return_periods():
     init_logger()
 
+    # ensure you have enough information provided to return a response
     if request.args.get('reach_id', '') == '':
         if request.args.get('lat', '') == '' or request.args.get('lon', '') == '' or \
                 request.args.get('region', '') == '':
@@ -170,9 +161,7 @@ def return_periods():
             return jsonify({"error": "Either reach_id or lat and lon parameters are required as input."}), 422
 
     try:
-        # Call the service
         return return_periods_handler(request)
-
     except:
         print(sys.exc_info()[0])
         logging.exception(sys.exc_info()[0])
@@ -185,6 +174,7 @@ def return_periods():
 def seasonal_average():
     init_logger()
 
+    # ensure you have enough information provided to return a response
     if request.args.get('reach_id', '') == '':
         if request.args.get('lat', '') == '' or request.args.get('lon', '') == '' or \
                 request.args.get('region', '') == '':
@@ -192,9 +182,7 @@ def seasonal_average():
             return jsonify({"error": "Either reach_id or lat and lon parameters are required as input."}), 422
 
     try:
-        # Call the service
         return seasonal_average_handler(request)
-
     except:
         print(sys.exc_info()[0])
         logging.exception(sys.exc_info()[0])
@@ -208,9 +196,7 @@ def available_data():
     init_logger()
 
     try:
-        # Call the service
         return get_available_data_handler()
-
     except:
         print(sys.exc_info()[0])
         logging.exception(sys.exc_info()[0])
@@ -241,7 +227,7 @@ def dates():
 
     try:
         # Call the service
-        return get_dates_handler(request)
+        return available_dates_handler(request)
 
     except:
         print(sys.exc_info()[0])
