+++ azcopy/file_mount.json
@@ -14,7 +14,7 @@
     {
         "accountName":"globalfloodsdiag360",
         "sasSig":"o8u%2BoutumDqKQL4A1I2Yft9T0S5M%2FJRBlMPuptfVehc%3D",
-        "volumeName":"era-5-rapid",
+        "volumeName":"era-5",
         "mappedDirectory":"/mnt/output/era-5"
     },
     {
